requirements.txt
# ScaleTent Requirements
# Core dependencies
numpy>=1.22.0
opencv-python>=4.7.0
opencv-contrib-python>=4.7.0
pytorchcv>=0.0.67
pytorch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8
onnx>=1.14.0
onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0; platform_system != "Darwin" or platform_machine != "arm64"

# API and Web dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
websockets>=11.0.0
jinja2>=3.1.2
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.1.0

# Database dependencies
pymongo>=4.3.3
motor>=3.1.2
redis>=4.5.5

# Utilities
pyyaml>=6.0
pillow>=9.5.0
scipy>=1.10.1
matplotlib>=3.7.1
pandas>=2.0.0
tqdm>=4.65.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Optional dependencies
mediapipe>=0.10.0; platform_system != "Linux" or platform_machine != "aarch64"  # For face detection
facenet-pytorch>=2.5.2  # For face recognition
insightface>=0.7.0; platform_system != "Darwin" or platform_machine != "arm64"  # Alternative face recognition
tensorrt>=8.6.0; platform_system == "Linux" and platform_machine != "aarch64"  # For TensorRT acceleration

# Development dependencies
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0
pytest-cov>=4.1.0